server/pyproject.toml
[project]
name = "grok-api-server"
version = "0.1.0"
description = "FastMCP server for Grok API (xAI) with web search grounding"
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
